+++ src/graphics/mesh.cpp
@@ -12,8 +12,8 @@ VkVertexInputBindingDescription Vertex::getBindingDescription() {
     return bindingDescription;
 }
 
-std::array<VkVertexInputAttributeDescription, 5> Vertex::getAttributeDescriptions() {
-    std::array<VkVertexInputAttributeDescription, 5> attributeDescriptions{};
+std::array<VkVertexInputAttributeDescription, 6> Vertex::getAttributeDescriptions() {
+    std::array<VkVertexInputAttributeDescription, 6> attributeDescriptions{};
 
     attributeDescriptions[0].binding = 0;
     attributeDescriptions[0].location = 0;
@@ -39,6 +39,11 @@ std::array<VkVertexInputAttributeDescription, 5> Vertex::getAttributeDescription
     attributeDescriptions[4].location = 4;
     attributeDescriptions[4].format = VK_FORMAT_R32_SFLOAT;
     attributeDescriptions[4].offset = offsetof(Vertex, auxiliary);
+
+    attributeDescriptions[5].binding = 0;
+    attributeDescriptions[5].location = 5;
+    attributeDescriptions[5].format = VK_FORMAT_R32_SFLOAT;
+    attributeDescriptions[5].offset = offsetof(Vertex, soilId);
 
     return attributeDescriptions;
 }